# Optimize several predictor transforms