Before:
```
test lossless_transform::benches::predictor01 ... bench:         493.49 ns/iter (+/- 3.76) = 2068 MB/s
test lossless_transform::benches::predictor05 ... bench:       1,154.43 ns/iter (+/- 17.28) = 883 MB/s
test lossless_transform::benches::predictor06 ... bench:         722.79 ns/iter (+/- 10.92) = 1412 MB/s
test lossless_transform::benches::predictor07 ... bench:         796.46 ns/iter (+/- 18.21) = 1281 MB/s
```

After:
```
test lossless_transform::benches::predictor01 ... bench:         197.49 ns/iter (+/- 1.53) = 5177 MB/s
test lossless_transform::benches::predictor05 ... bench:         460.01 ns/iter (+/- 1.27) = 2217 MB/s
test lossless_transform::benches::predictor06 ... bench:         394.44 ns/iter (+/- 19.60) = 2588 MB/s
test lossless_transform::benches::predictor07 ... bench:         445.22 ns/iter (+/- 6.88) = 2292 MB/s
```

Related to #94